+++ order/src/test/java/com/food/order/payment/PaymentDoTest.java
@@ -1,4 +1,4 @@
-package com.food.order;
+package com.food.order.payment;
 
 import com.food.common.order.business.internal.dto.OrderDto;
 import com.food.common.payment.business.internal.model.PaymentLogDto;
@@ -7,14 +7,14 @@ import com.food.order.error.DuplicatedPaymentException;
 import com.food.order.error.InvalidPaymentException;
 import com.food.order.error.NotFoundOrderException;
 import com.food.order.error.PaymentErrors;
-import com.food.order.mock.MockOrder;
-import com.food.order.mock.MockPayment;
-import com.food.order.mock.MockRequestUser;
+import com.food.order.payment.mock.MockOrder;
+import com.food.order.payment.mock.MockPayment;
+import com.food.order.payment.mock.MockRequestUser;
 import com.food.common.payment.business.external.model.PaymentDoRequest;
-import com.food.order.stubrepository.StubOrderService;
-import com.food.order.stubrepository.StubPaymentLogService;
-import com.food.order.stubrepository.StubPaymentService;
-import com.food.order.stubrepository.StubPointService;
+import com.food.order.payment.stubrepository.StubOrderService;
+import com.food.order.payment.stubrepository.StubPaymentLogService;
+import com.food.order.payment.stubrepository.StubPaymentService;
+import com.food.order.payment.stubrepository.StubPointService;
 import com.food.order.business.DefaultPayService;
 import com.food.common.payment.business.external.PayService;
 import org.junit.jupiter.api.BeforeEach;
