+++ order/src/test/java/com/food/order/payment/mock/MockPayment.java
@@ -1,4 +1,4 @@
-package com.food.order.mock;
+package com.food.order.payment.mock;
 
 import com.food.common.order.business.internal.dto.OrderDto;
 import com.food.common.payment.enumeration.PaymentActionType;
